Implement delete check for Products(LinkText,Type)

+++ Mvc5Memberships/Areas/Admin/Controllers/ProductLinkTextController.cs
@@ -112,8 +112,14 @@ namespace Mvc5Memberships.Areas.Admin.Controllers
         public async Task<ActionResult> DeleteConfirmed(int id)
         {
             ProductLinkText productLinkText = await db.ProductLinkTexts.FindAsync(id);
-            db.ProductLinkTexts.Remove(productLinkText);
-            await db.SaveChangesAsync();
+            var isUnused = await db.Products.CountAsync(p => p.ProductLinkTextId == id) == 0;
+
+            if (isUnused)
+            {
+                db.ProductLinkTexts.Remove(productLinkText);
+                await db.SaveChangesAsync();
+            }
+
             return RedirectToAction("Index");
         }
